use index instead of media id as key in PersonRow to prevent crash
