+++ version.properties
@@ -1,2 +1,2 @@
-#Fri Feb 03 21:55:21 IST 2023
-patch=12
+#Sun Feb 05 00:26:49 IST 2023
+patch=13
